Create action creator for fetchcartdata and displays it
