+++ 11-advanced-redux/src/App.js
@@ -4,7 +4,7 @@ import Cart from './components/Cart/Cart';
 import Layout from './components/Layout/Layout';
 import Products from './components/Shop/Products';
 import Notification from './components/UI/Notification';
-import { sendCartData } from './store/cartSlice';
+import { fetchCartData, sendCartData } from './store/cartAction';
 let isInitial = true;
 function App() {
     const dispatch = useDispatch();
@@ -12,10 +12,15 @@ function App() {
     const cart = useSelector((state) => state.cart);
     const notification = useSelector((state) => state.ui.notification);
     useEffect(() => {
-        dispatch(sendCartData(cart))
+        dispatch(fetchCartData());
+    }, [dispatch]);
+    useEffect(() => {
         if (isInitial) {
-          isInitial = false;
-          return
+            isInitial = false;
+            return;
+        }
+        if (cart.changed) {
+          dispatch(sendCartData(cart));
         }
     }, [cart, dispatch]);
