+++ 11-advanced-redux/src/store/cartSlice.js
@@ -1,11 +1,11 @@
 import { createSlice } from '@reduxjs/toolkit';
-import { uiAction } from './uiSlice';
 
 const cartSlice = createSlice({
     name: 'cart',
     initialState: {
         items: [],
         totalQuantity: 0,
+        changed: false,
     },
     reducers: {
         replaceCart(state, action) {
@@ -18,6 +18,7 @@ const cartSlice = createSlice({
                 (item) => item.id === newItem.id
             );
             state.totalQuantity++;
+            state.changed = true;
             if (!existingItem) {
                 state.items.push({
                     id: newItem.id,
@@ -36,60 +37,17 @@ const cartSlice = createSlice({
             const id = action.payload;
             const existingItem = state.items.find((item) => item.id === id);
             state.totalQuantity--;
+            state.changed = true;
             if (existingItem.quantity === 1) {
                 state.items = state.items.filter((item) => item.id !== id);
             } else {
                 existingItem.quantity--;
+                existingItem.totalPrice =
+                    existingItem.totalPrice - existingItem.price;
             }
         },
     },
 });
-
-export const sendCartData = (cart) => {
-    return async (dispatch) => {
-        dispatch(
-            uiAction.showNotification({
-                status: 'pending',
-                title: 'Sending...',
-                message: 'Sending cart data!',
-            })
-        );
-
-        const sendRequest = async () => {
-            const response = await fetch(
-                'https://react-http-60ae9-default-rtdb.firebaseio.com/cart.json',
-                {
-                    method: 'PUT',
-                    body: JSON.stringify(cart),
-                }
-            );
-
-            if (!response.ok) {
-                throw new Error('Sending cart data failed.');
-            }
-        };
-
-        try {
-            await sendRequest();
-
-            dispatch(
-                uiAction.showNotification({
-                    status: 'success',
-                    title: 'Success!',
-                    message: 'Sent cart data successfully!',
-                })
-            );
-        } catch (error) {
-            dispatch(
-                uiAction.showNotification({
-                    status: 'error',
-                    title: 'Error!',
-                    message: 'Sending cart data failed!',
-                })
-            );
-        }
-    };
-};
 
 export const cartActions = cartSlice.actions;
